Break out fdfit into separate files/folders
