+++ lumicks/pylake/__init__.py
@@ -2,8 +2,8 @@ from .__about__ import (__author__, __copyright__, __doc__, __email__, __license
                         __title__, __url__, __version__)
 
 from .file import *
-from .fdmodels import force_model
-from .fdfit import FitObject
+from .fitting.fitobject import FitObject
+from .fitting.fdmodels import force_model
 
 
 def pytest(args=None, plugins=None):
