+++ lumicks/pylake/fitting/fdmodels.py
@@ -1,7 +1,8 @@
-from lumicks.pylake.fdfit import Model, Parameter, invert_function, invert_jacobian, InverseModel
+from .model import Model
+from .parameters import Parameter
+from .detail.derivative_manipulation import invert_function, invert_jacobian
+from .model import Model, InverseModel
 import numpy as np
-import inspect
-import scipy.optimize as optim
 
 
 def force_model(name, model_type):
